pyproject.toml
[project]
name = "remi-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aerich>=0.9.0",
    "aiofile>=3.9.0",
    "bcrypt>=4.3.0",
    "casbin>=1.43.0",
    "fastapi>=0.115.12",
    "loguru>=0.7.3",
    "maimai-py>=1.0.6",
    "passlib>=1.7.4",
    "pydantic[email]>=2.11.4",
    "tortoise-orm[aiomysql]>=0.25.0",
    "uvicorn>=0.34.2",
    "wahlap-mai-ass-expander",
]

[tool.aerich]
tortoise_orm = "cfgs.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[tool.uv.sources]
wahlap-mai-ass-expander = { path = "../Wahlap-Mai-Ass-Expander", editable = true }
